Added deleteMessages and events from user, message classes

+++ src/messages.ts
@@ -1,26 +1,29 @@
-import { UUID } from "crypto";
+import { UUID, privateDecrypt, publicDecrypt } from "crypto";
 
 import { HTTPConnection, SocketConnection } from "./connections.js";
 import * as types from "./types/messages.js";
+import EventEmitter from "events";
 
-export class Message {
+export class Message extends EventEmitter {
     readonly id: string;
     readonly author: UUID;
     readonly receiver: string;
     public content: string;
     readonly datetime: Date;
     readonly editDatetime: Date | null;
-    readonly read: boolean;
+    public read: boolean;
     readonly connection: HTTPConnection | undefined;
     readonly socket: SocketConnection | undefined;
     
     public encrypted: boolean;
 
     constructor(data: types.Message, connection: HTTPConnection | undefined = undefined, socket: SocketConnection | undefined = undefined) {
+        super();
+        
         this.connection = connection
         this.socket = socket
         
-        this.id = data._id
+        this.id = data.id
         this.author = data.author
         this.receiver = data.receiver
         this.content = data.content
@@ -29,10 +32,57 @@ export class Message {
         this.read = data.read
 
         this.encrypted = true;
+
+
+        this.socket?.on("messageEdit", (id: string, content: string) => {
+            if (id !== this.id)
+                return;
+
+            
+            if (!this.connection) {
+                this.content = content
+                this.encrypted = true
+    
+                return;
+            }
+
+            let key = this.connection.keys[this.author].private
+            if (key) {
+                this.content = privateDecrypt(key, Buffer.from(content)).toString("utf-8");
+                
+                this.encrypted = false                
+            }
+
+            this.emit("messageEdit", ...Object.values(arguments));
+        });
+
+        this.socket?.on("messageDelete", (id: string) => {
+            if (id !== this.id)
+                return;
+
+            this.deleteMessage = async () => {return false};
+            this.editMessage = async () => {return false};
+            this.markAsRead = async () => {return false};
+
+            this.emit("messageDelete", ...Object.values(arguments));
+        });
+
+        this.socket?.on("readMessage", (id: string) => {
+            if (id !== this.id)
+                return;
+            
+            this.read = true;
+
+            this.markAsRead = async () => {return false};
+
+            this.emit("readMessage", ...Object.values(arguments));
+        });
     }
 
     async deleteMessage() : Promise<boolean> {
         if (!this.connection) return false;
+
+        if (this.connection.account.uuid !== this.author) return false;
 
         await this.connection.deleteMessage(this.id);
 
@@ -41,6 +91,8 @@ export class Message {
 
     async editMessage(content: string) : Promise<boolean> {
         if (!this.connection) return false;
+
+        if (this.connection.account.uuid !== this.author) return false;
 
         await this.connection.editMessage(this.id, content);
         
@@ -51,6 +103,8 @@ export class Message {
 
     async markAsRead() : Promise<boolean> {
         if (!this.socket || this.read) return false;
+
+        if (this.socket.account.uuid !== this.author) return false;
 
         await this.socket.markMessageAsRead(this.id);
